Refactor: query into a separate hook
