+++ admin-ui/src/component/my-units/MyUnitView.tsx
@@ -2,19 +2,17 @@ import React from "react";
 import { H1 } from "common/src/common/typography";
 import { useTranslation } from "react-i18next";
 import { useParams } from "react-router-dom";
-import { LocationType, Query, QueryUnitsArgs } from "common/types/gql-types";
+import { LocationType } from "common/types/gql-types";
 import { LoadingSpinner } from "hds-react";
-import { useQuery } from "@apollo/client";
 import { publicUrl } from "../../common/const";
 import { parseAddress } from "../../common/util";
-import { useNotification } from "../../context/NotificationContext";
 import { Container } from "../../styles/layout";
 import BreadcrumbWrapper from "../BreadcrumbWrapper";
 import withMainMenu from "../withMainMenu";
-import { UNIT_QUERY } from "./queries";
 import ReservationUnitCalendarView from "./ReservationUnitCalendarView";
 import UnitReservationsView from "./UnitReservationsView";
 import { TabHeader, TabPanel, Tabs } from "../Tabs";
+import { useUnitQuery } from "./hooks";
 
 type Params = {
   unitId: string;
@@ -22,7 +20,6 @@ type Params = {
 };
 
 const MyUnitView = () => {
-  const { notifyError } = useNotification();
   const { unitId } = useParams<Params>();
   const { t } = useTranslation();
 
@@ -34,16 +31,7 @@ const MyUnitView = () => {
     { key: "reservation-unit", label: `${t("MyUnits.Calendar.Tabs.byUnit")}` },
   ];
 
-  const { loading, data: unitData } = useQuery<Query, QueryUnitsArgs>(
-    UNIT_QUERY,
-    {
-      skip: unitId == null,
-      onError: (err) => {
-        notifyError(err.message);
-      },
-      variables: { pk: [unitId ?? ""], offset: 0 },
-    }
-  );
+  const { loading, data: unitData } = useUnitQuery(unitId);
 
   const unit = unitData?.units?.edges[0];
