+++ admin-ui/src/component/my-units/hooks/queries.ts
@@ -29,3 +29,30 @@ export const OPTIONS_QUERY = gql`
     }
   }
 `;
+
+export const UNIT_QUERY = gql`
+  query units($pk: [ID]) {
+    units(pk: $pk) {
+      edges {
+        node {
+          location {
+            addressStreetFi
+            addressZip
+            addressCityFi
+          }
+          nameFi
+          pk
+          serviceSectors {
+            nameFi
+          }
+          reservationUnits {
+            pk
+            spaces {
+              pk
+            }
+          }
+        }
+      }
+    }
+  }
+`;
